CCMSPUI-458 Updated build.gradle to include up to date versions of SOA and EBS with EBS endpoint migrated
