+++ build.gradle
@@ -37,9 +37,9 @@ repositories {
 
 dependencies {
 
-	implementation 'uk.gov.laa.ccms.data:data-api:0.0.29'
+	implementation 'uk.gov.laa.ccms.data:data-api:0.0.30-bc6420d-SNAPSHOT'
 
-	implementation 'uk.gov.laa.ccms.soa.gateway:soa-gateway-api:0.0.46'
+	implementation 'uk.gov.laa.ccms.soa.gateway:soa-gateway-api:0.0.46-17d5bbf-SNAPSHOT'
 
 	implementation 'uk.gov.laa.ccms.caab:caab-api:0.0.40'
