+++ src/test/java/uk/gov/laa/ccms/caab/mapper/ApplicationMapperTest.java
@@ -16,10 +16,9 @@ import static uk.gov.laa.ccms.caab.constants.ApplicationConstants.AWARD_TYPE_OTH
 import static uk.gov.laa.ccms.caab.constants.ApplicationConstants.OPPONENT_TYPE_INDIVIDUAL;
 import static uk.gov.laa.ccms.caab.constants.ApplicationConstants.OPPONENT_TYPE_ORGANISATION;
 import static uk.gov.laa.ccms.caab.constants.ApplicationConstants.STATUS_DRAFT;
-import static uk.gov.laa.ccms.caab.util.AssessmentModelUtils.buildProceedingNameAttribute;
 import static uk.gov.laa.ccms.caab.util.CaabModelUtils.buildApplicationDetail;
 import static uk.gov.laa.ccms.caab.util.CaabModelUtils.buildBaseApplication;
-import static uk.gov.laa.ccms.caab.util.CaabModelUtils.buildOpponent;
+import static uk.gov.laa.ccms.caab.util.EbsModelUtils.buildCaseSummary;
 import static uk.gov.laa.ccms.caab.util.EbsModelUtils.buildPriorAuthorityDetail;
 import static uk.gov.laa.ccms.caab.util.EbsModelUtils.buildPriorAuthorityTypeDetail;
 import static uk.gov.laa.ccms.caab.util.EbsModelUtils.buildUserDetail;
@@ -27,7 +26,6 @@ import static uk.gov.laa.ccms.caab.util.SoaModelUtils.buildAddressDetail;
 import static uk.gov.laa.ccms.caab.util.SoaModelUtils.buildAssessmentResult;
 import static uk.gov.laa.ccms.caab.util.SoaModelUtils.buildBaseClient;
 import static uk.gov.laa.ccms.caab.util.SoaModelUtils.buildCaseDetail;
-import static uk.gov.laa.ccms.caab.util.SoaModelUtils.buildCaseSummary;
 import static uk.gov.laa.ccms.caab.util.SoaModelUtils.buildCostAward;
 import static uk.gov.laa.ccms.caab.util.SoaModelUtils.buildCostLimitation;
 import static uk.gov.laa.ccms.caab.util.SoaModelUtils.buildFinancialAward;
@@ -45,8 +43,6 @@ import static uk.gov.laa.ccms.caab.util.SoaModelUtils.buildTimeRelatedAward;
 import java.math.BigDecimal;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
-import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
 import java.util.Collections;
 import java.util.Date;
 import java.util.List;
@@ -91,6 +87,7 @@ import uk.gov.laa.ccms.caab.model.ReferenceDataItemDetail;
 import uk.gov.laa.ccms.caab.model.ScopeLimitationDetail;
 import uk.gov.laa.ccms.caab.model.StringDisplayValue;
 import uk.gov.laa.ccms.caab.model.TimeRecoveryDetail;
+import uk.gov.laa.ccms.data.model.CaseSummary;
 import uk.gov.laa.ccms.data.model.CommonLookupValueDetail;
 import uk.gov.laa.ccms.data.model.OfficeDetail;
 import uk.gov.laa.ccms.data.model.OutcomeResultLookupValueDetail;
@@ -101,11 +98,9 @@ import uk.gov.laa.ccms.soa.gateway.model.Award;
 import uk.gov.laa.ccms.soa.gateway.model.BaseClient;
 import uk.gov.laa.ccms.soa.gateway.model.CaseDetail;
 import uk.gov.laa.ccms.soa.gateway.model.CaseDoc;
-import uk.gov.laa.ccms.soa.gateway.model.CaseSummary;
 import uk.gov.laa.ccms.soa.gateway.model.CategoryOfLaw;
 import uk.gov.laa.ccms.soa.gateway.model.CostLimitation;
 import uk.gov.laa.ccms.soa.gateway.model.LinkedCase;
-import uk.gov.laa.ccms.soa.gateway.model.NameDetail;
 import uk.gov.laa.ccms.soa.gateway.model.OpaAttribute;
 import uk.gov.laa.ccms.soa.gateway.model.OtherParty;
 import uk.gov.laa.ccms.soa.gateway.model.OtherPartyOrganisation;
