+++ src/test/java/uk/gov/laa/ccms/caab/util/PaginationUtilTest.java
@@ -12,8 +12,8 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import uk.gov.laa.ccms.caab.exception.CaabApplicationException;
+import uk.gov.laa.ccms.data.model.CaseSummary;
 import uk.gov.laa.ccms.data.model.Notification;
-import uk.gov.laa.ccms.soa.gateway.model.CaseSummary;
 
 class PaginationUtilTest {
