+++ src/test/java/uk/gov/laa/ccms/caab/util/SoaModelUtils.java
@@ -19,7 +19,6 @@ import uk.gov.laa.ccms.soa.gateway.model.BaseContact;
 import uk.gov.laa.ccms.soa.gateway.model.CaseDetail;
 import uk.gov.laa.ccms.soa.gateway.model.CaseDoc;
 import uk.gov.laa.ccms.soa.gateway.model.CaseStatus;
-import uk.gov.laa.ccms.soa.gateway.model.CaseSummary;
 import uk.gov.laa.ccms.soa.gateway.model.CategoryOfLaw;
 import uk.gov.laa.ccms.soa.gateway.model.ClientDetail;
 import uk.gov.laa.ccms.soa.gateway.model.ClientDetailDetails;
@@ -588,16 +587,6 @@ public class SoaModelUtils {
         .remainderAuthorisation(remainderAuth)
         .contractualDevolvedPowers("CATDEVPOW")
         .authorisationType("AUTHTYPE1");
-  }
-
-  public static CaseSummary buildCaseSummary() {
-    return new CaseSummary()
-        .caseReferenceNumber("caseref")
-        .caseStatusDisplay("the status")
-        .categoryOfLaw("CAT")
-        .client(buildBaseClient())
-        .feeEarnerName("feeearner")
-        .providerCaseReferenceNumber("prov");
   }
 
   public static BaseContact buildBaseContact() {
